fix: use fp16 instead of int8 quantization for embedding model
int8 severely degrades cross-lingual and multilingual matching quality
in bge-m3. fp16 gives much better accuracy with reasonable memory use.
Requires reindex to take effect.

+++ src/embeddings/model.ts
@@ -17,7 +17,7 @@ async function getPipeline(modelId: string) {
   pipelinePromise = (async () => {
     try {
       const { pipeline } = await import('@huggingface/transformers');
-      pipelineInstance = await pipeline('feature-extraction', modelId, { dtype: 'int8' });
+      pipelineInstance = await pipeline('feature-extraction', modelId, { dtype: 'fp16' });
       currentModelId = modelId;
       return pipelineInstance;
     } catch (e) {
